refactor: shrink hot runtime structs to u8 fields
Narrow small-value fields from u32/usize/i32 to u8 across core state,
snapshot, and display structs:
- types.rs: Allocation {target_idx, source_slot, color}, MemberContribution.slot,
  PositionChangeEvent.old/new_position, AbilityApplication.heart_color,
  Adjustment.color, AbilityBonus.color, SurplusHeart.old_value (u64→u8)
- GameState: RPS choices + pending_rps_player_id i32→u8
- ability_queue/resolver signatures: RPS handlers take u8

All serde-compatible (integers serialize identically); full test suite green.

+++ engine/src/ability/condition/state.rs
@@ -441,7 +441,9 @@ impl<'a> ConditionContext<'a> {
                 } else {
                     event.new_position
                 };
-                if check_pos >= pos_names.len() || pos_names[check_pos] != req_pos {
+                if (check_pos as usize) >= pos_names.len()
+                    || pos_names[check_pos as usize] != req_pos
+                {
                     return false;
                 }
             }
